pyproject.toml
[tool.poetry]
name = "gcp-monitoring"
version = "0.1.0"
description = ""
authors = ["arbilgin <arda.bilgin@blok-z.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-monitoring = "^2.18.0"
flask = "^3.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
